add profile achievements fragment
