+++ app/src/main/java/org/stepik/android/view/profile/ui/fragment/ProfileFragment.kt
@@ -41,6 +41,7 @@ import org.stepik.android.presentation.profile.ProfileView
 import org.stepik.android.view.base.ui.span.TypefaceSpanCompat
 import org.stepik.android.view.injection.profile.ProfileComponent
 import org.stepik.android.view.profile.ui.activity.ProfileActivity
+import org.stepik.android.view.profile_achievements.ui.fragment.ProfileAchievementsFragment
 import org.stepik.android.view.profile_detail.ui.fragment.ProfileDetailFragment
 import org.stepik.android.view.ui.delegate.ViewStateDelegate
 import ru.nobird.android.view.base.ui.extension.argument
@@ -145,6 +146,7 @@ class ProfileFragment : Fragment(), ProfileView {
         if (savedInstanceState == null) {
             childFragmentManager.commitNow {
                 add(R.id.container, ProfileDetailFragment.newInstance(userId))
+                add(R.id.container, ProfileAchievementsFragment.newInstance(userId))
             }
         }
