+++ app/src/main/res/values/dimens.xml
@@ -265,7 +265,7 @@
     <!-- Profile -->
     <dimen name="profile_header_elevation">4dp</dimen>
     <dimen name="profile_header_margin">16dp</dimen>
-    <dimen name="profile_header_stats_margin">8dp</dimen
+    <dimen name="profile_header_stats_margin">8dp</dimen>
 
     <dimen name="profile_block_margin">24dp</dimen>
     <dimen name="profile_block_vertical_margin">16dp</dimen>
